// src/types/product.ts
export type TMedicine = {
  brand: string;
  category: string;
  description: string;
  dosage: string;
  expiry_date: string;
  form: string;
  image: string;
  name: string;
  prescription_required: boolean;
  price: number;
  side_effects: string[];
  stock: number;
  uses: string[];
  _id: string;
};